Bring code formatting into compliance with new code style
Although every effort was made for it to be seamless, the switch from the use of Artistic Style to ClangFormat tools for
code formatting in Arduino IDE 2.x resulted in some changes to the standard Arduino code style. The necessary changes to
bring the code into alignment with the new style are made here.